fixed build err

+++ src/app/Components/CardItems/CardDelete/CardDelete.tsx
@@ -1,36 +1,32 @@
 import axios from 'axios';
 import styles from './CardDelete.module.scss';
 import Image from 'next/image';
-import { useEffect, useState } from 'react';
 
 type Props = {
     onClick?: () => void;
     id?: number;
-}
+};
 
 const CardDelete = (props: Props) => {
-    // const [playlists, setPlaylists] = useState([]);
+    const PlaylistDelete = async () => {
+        if (props.id) { // Ensure there's an ID before attempting to delete
+            await axios.delete(`https://interstellar-1-pdzj.onrender.com/playlist/${props.id}`)
+                .then(() => {
+                    alert('Playlist has been deleted');
+                })
+                .catch((error) => {
+                    console.error('Error deleting playlist:', error);
+                });
+        } else {
+            alert('No playlist ID provided.');
+        }
+    };
 
-    // useEffect(() => {
-    //      axios.get(`https://interstellar-1-pdzj.onrender.com/playlist`)
-    //     .then((r) => {
-    //         setPlaylists(r.data)
-    //     })
-    // },[])
-
-    const PlaylistDelete = async (values: any) => {
-        await axios.delete(`https://interstellar-1-pdzj.onrender.com/playlist/${props.id}`)
-         .then(() => {
-             alert('Playlist has deleted')
-         })
-     }
-
-    return(
+    return (
         <div className={styles.container} onClick={PlaylistDelete}>
             <Image src={'/icon/card-delete-icon.svg'} alt={'image'} width={26} height={26} />
         </div>
-    )
-}
-
+    );
+};
 
 export default CardDelete;
